issue_43 - using proper names for modules and packages
