+++ sushibar-backend-app/scripts/startKafka.sh
