+++ sushibar-backend-app/src/main/java/org/virtualsushibar/backend/app/MainApplication.java
@@ -1,0 +1,15 @@
+package org.virtualsushibar.backend.app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.kafka.annotation.EnableKafka;
+
+@EnableKafka
+@SpringBootApplication
+public class MainApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MainApplication.class, args);
+	}
+
+}
