+++ sushibar-backend-app/src/main/java/org/virtualsushibar/backend/app/api/Meals.java
@@ -1,0 +1,7 @@
+package org.virtualsushibar.backend.app.api;
+
+public enum Meals {
+    SPAGHETTI,
+    PIZZA,
+    RICE
+}
