+++ sushibar-backend-app/src/main/java/org/virtualsushibar/backend/app/api/dto/Message.java
@@ -1,0 +1,14 @@
+package org.virtualsushibar.backend.app.api.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Builder
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Message {
+    private String mgs;
+}
