+++ sushibar-backend-app/src/main/java/org/virtualsushibar/backend/app/api/dto/OrderRequest.java
@@ -1,0 +1,16 @@
+package org.virtualsushibar.backend.app.api.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.virtualsushibar.backend.app.api.Meals;
+
+
+@Builder
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class OrderRequest {
+    private  Meals meal;
+}
