+++ sushibar-backend-app/src/main/java/org/virtualsushibar/backend/app/api/dto/OrderResponse.java
@@ -1,0 +1,10 @@
+package org.virtualsushibar.backend.app.api.dto;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class OrderResponse {
+    private String orderConfirmationID;
+}
