+++ sushibar-backend-app/src/main/java/org/virtualsushibar/backend/app/dao/repository/OrderRepository.java
@@ -1,0 +1,12 @@
+package org.virtualsushibar.backend.app.dao.repository;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.virtualsushibar.backend.app.dao.entity.OrderEntity;
+
+import java.util.Optional;
+
+public interface OrderRepository extends MongoRepository<OrderEntity, String> {
+
+    Optional<OrderEntity> findByMeal(String meal);
+
+}
