+++ sushibar-backend-app/src/main/java/org/virtualsushibar/backend/app/kafka/config/KafkaConsumerConfig.java
@@ -1,0 +1,56 @@
+package org.virtualsushibar.backend.app.kafka.config;
+
+
+import io.confluent.kafka.serializers.KafkaAvroDeserializer;
+import io.confluent.kafka.serializers.KafkaAvroDeserializerConfig;
+import lombok.RequiredArgsConstructor;
+import org.apache.kafka.clients.consumer.ConsumerConfig;
+import org.apache.kafka.common.serialization.IntegerDeserializer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
+import org.springframework.kafka.core.ConsumerFactory;
+import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
+import org.springframework.kafka.listener.ConcurrentMessageListenerContainer;
+import org.springframework.kafka.listener.ContainerProperties;
+import org.virtualsushibar.backend.avro.Order;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Configuration
+@RequiredArgsConstructor
+public class KafkaConsumerConfig {
+
+    private final AbstractKafkaConfigs configs;
+
+
+    @Bean
+    public ConsumerFactory<String, Order> consumerFactory() {
+        Map<String, Object> configProps = new HashMap<>();
+        configProps.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, configs.getBootstrapAddress());
+        configProps.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, IntegerDeserializer.class);
+        configProps.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, KafkaAvroDeserializer.class);
+        configProps.put(KafkaAvroDeserializerConfig.SPECIFIC_AVRO_READER_CONFIG, "true");
+        configProps.put(ConsumerConfig.REQUEST_TIMEOUT_MS_CONFIG, "20971520");
+        configProps.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
+        configProps.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG,false);
+        configProps.put("schema.registry.url", "http://localhost:8081");
+
+        return new DefaultKafkaConsumerFactory<>(configProps);
+    }
+
+
+    @Bean
+    public ConcurrentKafkaListenerContainerFactory<String, Order>
+    kafkaListenerContainerFactory(ConsumerFactory<String, Order> consumerFactory) {
+
+        ConcurrentKafkaListenerContainerFactory<String, Order> factory =
+                new ConcurrentKafkaListenerContainerFactory<>();
+        factory.setConsumerFactory(consumerFactory);
+        factory.setConcurrency(6);
+        factory.getContainerProperties().setAckMode(ContainerProperties.AckMode.MANUAL_IMMEDIATE);
+        return factory;
+    }
+
+}
