+++ sushibar-backend-app/src/test/java/org/virtualsushibar/backend/app/MainApplicationTests.java
@@ -1,0 +1,13 @@
+package org.virtualsushibar.backend.app;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MainApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
