--- sushibar-backend-kafka/src/main/java/org/virtualsushibar/backend/MainApplication.java
@@ -1,15 +1,0 @@
-package org.virtualsushibar.backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.kafka.annotation.EnableKafka;
-
-@EnableKafka
-@SpringBootApplication
-public class MainApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MainApplication.class, args);
-	}
-
-}
