--- sushibar-backend-kafka/src/main/java/org/virtualsushibar/backend/api/Meals.java
@@ -1,7 +1,0 @@
-package org.virtualsushibar.backend.api;
-
-public enum Meals {
-    SPAGHETTI,
-    PIZZA,
-    RICE
-}
