--- sushibar-backend-kafka/src/main/java/org/virtualsushibar/backend/api/dto/Message.java
@@ -1,14 +1,0 @@
-package org.virtualsushibar.backend.api.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Message {
-    private String mgs;
-}
