--- sushibar-backend-kafka/src/main/java/org/virtualsushibar/backend/api/dto/OrderResponse.java
@@ -1,10 +1,0 @@
-package org.virtualsushibar.backend.api.dto;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class OrderResponse {
-    private String orderConfirmationID;
-}
