--- sushibar-backend-kafka/src/main/java/org/virtualsushibar/backend/dao/entity/OrderEntity.java
@@ -1,16 +1,0 @@
-package org.virtualsushibar.backend.dao.entity;
-
-import lombok.Builder;
-import lombok.Data;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Data
-@Builder
-@Document
-public class OrderEntity {
-    @Id
-    public String id;
-    public String meal;
-    public int amount;
-}
