--- sushibar-backend-kafka/src/test/java/org/virtualsushibar/backend/MainApplicationTests.java
@@ -1,13 +1,0 @@
-package org.virtualsushibar.backend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MainApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
